Get publication metadata

Parse basic publication metadata from a decoded OpenAIRE Research Graph record.

oarg_publications_md(doc)

Arguments

doc
document of class `xml_document` from the OpenAIRE research graph publication subset

Value

The result is a tibble with each row representing a publication. Here are the returned columns and descriptions:

type	Research output type (literature publication, dataset, software)

title	Publication title

journal	Journal title

publisher	Publisher

date_of_acceptance	Date of acceptance for publication

best_access_right	Access level

embargo_enddate	Date when the publication becomes openly available

resource_type	Type of dataset

authors	List-column with author details including names, order and ORCID

pids	List-column with persistent identifiers (PIDS) referencing the publication, e.g. DOI, PMID or PMCID. PIDs are particularly useful when cross-linking to other scholarly data sources

collected_from	Name and identifier of metadata source

source_ids	metadata source ids like OAI identifier

Examples

if (FALSE) {
# sample file delivered with this package
dump_eg <- system.file("extdata", "multiple_projects.xml", package = "openairegraph")
# load xml file
my_record <- xml2::read_xml(dump_eg)
# parse
out <- openairegraph::oarg_publications_md(my_record)

# Detailed author infos
openairegraph::oarg_publications_md(out)$authors

# Obtain linked persistent identifiers (PID)
openairegraph::oarg_publications_md(my_record)$pids
}